Updated comments, moved input man into init
